# openAPI/AppDoc.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /med/timeslots:
    get:
      tags:
        - timeslot-controller
      operationId: getAllTimeslot
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCreateTimeslotDTO'
    put:
      tags:
        - timeslot-controller
      operationId: updateTimeslot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeslotDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TimeslotDTO'
...
  /med/timeslots/add:
    post:
      tags:
        - timeslot-controller
      operationId: createTimeslot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimeslotDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CreateTimeslotDTO'
...
  /med/timeslots/{id}:
    get:
      tags:
        - timeslot-controller
      operationId: getTimeslotById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CreateTimeslotDTO'
    delete:
      tags:
        - timeslot-controller
      operationId: deleteTimeslot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
...
components:
  schemas:
...
    TimeslotDTO:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        date:
          type: string
          format: date-time
        office:
          type: integer
          format: int32
        timeslotId:
          type: integer
          format: int32
...
    CreateTimeslotDTO:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        date:
          type: string
          format: date-time
        office:
          type: integer
          format: int32
...
    Pageable:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    PageCreateTimeslotDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/CreateTimeslotDTO'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
...