# src/main/java/com/itrex/java/lab/dto/CreateTimeslotDTO.java
package com.itrex.java.lab.dto;

import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.AllArgsConstructor;
import lombok.experimental.SuperBuilder;

import java.time.Instant;

@Data
@SuperBuilder
@NoArgsConstructor
@AllArgsConstructor
public class CreateTimeslotDTO {

    private Instant startTime;
    private Instant date;
    private Integer office;

}